bloglist expansion, step10

+++ part4/bloglist/controllers/blogs.js
@@ -3,6 +3,14 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
+blogsRouter.get('/', async(request, response) => {
+    const blogs = await Blog
+        .find({})
+        .populate('user', { username: 1, name: 1 })
+
+    response.json(blogs.map(blog => blog.toJSON()))
+})
+
 blogsRouter.post('/', async(request, response) => {
     const body = request.body
 
@@ -29,17 +37,18 @@ blogsRouter.post('/', async(request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.get('/', async(request, response) => {
-    const blogs = await Blog
-        .find({})
-        .populate('user', { username: 1, name: 1 })
-
-    response.json(blogs.map(blog => blog.toJSON()))
-})
-
 blogsRouter.delete('/:id', async(request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === decodedToken.id) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(400).json({ error: 'invalid delete operation' })
+    }
 })
 
 blogsRouter.put('/:id', async(request, response) => {
